refactor: logic should only reflect five elements
Instead of focusing on all the elements of the birth chart, we are only
focusing on a subset of them. That is why the logic should take those
into account, and leave the rest as optional. If in the future we decide
to use those, we will have the logic ready.